@@ -1,5 +1,3 @@
-module github.com/creack/pty
+module github.com/creack/pty/v2
 
-go 1.13
-
-require golang.org/x/sys v0.0.0-20220721230656-c6bc011c0c49
+go 1.18
